管理者チェックの追加

+++ src/main/java/actions/EmployeeAction.java
@@ -41,6 +41,8 @@ public class EmployeeAction  extends ActionBase{
      */
     public void index() throws ServletException, IOException {
 
+        //管理者かどうかのチェック
+        if(checkAdmin()) {
         //指定されたページ数の一覧画面に表示するデータを取得
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
@@ -62,6 +64,7 @@ public class EmployeeAction  extends ActionBase{
 
         //一覧画面を表示
         forward(ForwardConst.FW_EMP_INDEX);
+        }
     }
 
     /**
@@ -71,22 +74,25 @@ public class EmployeeAction  extends ActionBase{
      */
     public void entryNew() throws ServletException, IOException {
 
+        //管理者かどうかのチェック
+        if(checkAdmin()) {
         putRequestScope(AttributeConst.TOKEN, getTokenId());  //CSRF対策用トークン
         putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());  //空の従業員インスタンス
 
         forward(ForwardConst.FW_EMP_NEW);
+        }
     }
 
     /**
      * 新規登録を行う
      * @throws IOException
      * @throws ServletException
-     * @thorws IOException
+     * @throws IOException
      */
     public void create() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if(checkToken()) {
+        if(checkAdmin() && checkToken()) {
 
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
@@ -135,6 +141,8 @@ public class EmployeeAction  extends ActionBase{
      */
     public void show() throws ServletException, IOException {
 
+        //管理者かどうかのチェック
+        if(checkAdmin()) {
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -149,6 +157,7 @@ public class EmployeeAction  extends ActionBase{
 
         //詳細画面の表示
         forward(ForwardConst.FW_EMP_SHOW);
+        }
     }
 
     /**
@@ -158,6 +167,8 @@ public class EmployeeAction  extends ActionBase{
      */
     public void edit() throws ServletException, IOException {
 
+        //管理者かどうかのチェック
+        if(checkAdmin()) {
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -173,6 +184,7 @@ public class EmployeeAction  extends ActionBase{
 
         //編集画面を表示する
         forward(ForwardConst.FW_EMP_EDIT);
+        }
     }
 
     /**
@@ -183,7 +195,7 @@ public class EmployeeAction  extends ActionBase{
     public void update() throws ServletException, IOException {
 
         //CSRF対策　tokenのチェック
-        if(checkToken()) {
+        if(checkAdmin() && checkToken()) {
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -230,7 +242,7 @@ public class EmployeeAction  extends ActionBase{
     public void destroy() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if(checkToken()) {
+        if(checkAdmin() && checkToken()) {
 
             //idを条件に従業員データを論理削除する
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -243,4 +255,29 @@ public class EmployeeAction  extends ActionBase{
         }
     }
 
+    /**
+     * ログイン中の従業員が管理者かどうかチェックし、管理者でなければエラー画面を表示
+     * true: 管理者 false: 管理者ではない
+     * @throws ServletException
+     * @throws IOException
+     */
+    private boolean checkAdmin() throws ServletException, IOException {
+
+        //セッションからログイン中ノン従業員情報を取得
+        EmployeeView ev = (EmployeeView)getSessionScope(AttributeConst.LOGIN_EMP);
+
+        //管理者でなければエラー画面を表示
+        if(ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+
+        }else {
+
+            return true;
+        }
+
+    }
+
 }
+
